requirements/codestyle.txt
black==20.8b1
flake8==3.8.4
flake8-isort==4.0.0
flake8-bugbear==20.11.1
flake8-builtins==1.5.3
flake8-comprehensions==3.3.1
flake8-mutable==1.2.0
flake8-pep3101==1.3.0
flake8-print==4.0.0
flake8-quotes==3.2.0
flake8-logging-format==0.6.0
flake8-debugger==4.0.0
flake8-django==1.1.1
